requirements.txt
PyPDF2==3.0.1
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
streamlit==1.26.0
langchain
google-generativeai
python-docx
faiss-cpu
python-dotenv
